feat: adds error handler
